Removed ptsa.filt; moved ptsa.filt.buttfilt into ptsa.data.filters.ButterworthFilter
